# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI 文章创作平台",
    "description": "基于多智能体编排的 AI 文章生成平台",
    "version": "0.0.1"
  },
  "paths": {
...
    "/user/register": {
      "post": {
        "tags": [
          "用户管理"
        ],
        "summary": "Register",
        "description": "用户注册（同一 IP 在窗口内限次）",
        "operationId": "register_api_user_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_int_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": [
          "用户管理"
        ],
        "summary": "Login",
        "description": "用户登录（单账号失败超限后锁定 + IP 级失败限流，防密码爆破与撞库）",
        "operationId": "login_api_user_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_LoginUserVO_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/logout": {
      "post": {
        "tags": [
          "用户管理"
        ],
        "summary": "Logout",
        "description": "用户登出（清除 Cookie 与 Redis 服务端会话）",
        "operationId": "logout_api_user_logout_post",
        "parameters": [
          {
            "name": "SESSION",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_bool_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/get/login": {
      "get": {
        "tags": [
          "用户管理"
        ],
        "summary": "Get Login User",
        "description": "获取当前登录用户",
        "operationId": "get_login_user_api_user_get_login_get",
        "parameters": [
          {
            "name": "SESSION",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_LoginUserVO_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/get": {
      "get": {
        "tags": [
          "用户管理"
        ],
        "summary": "Get User By Id",
        "description": "根据 ID 获取用户",
        "operationId": "get_user_by_id_api_user_get_get",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_UserVO_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/profile": {
      "get": {
        "tags": [
          "用户管理"
        ],
        "summary": "Get User Profile",
        "description": "获取当前登录用户的主页信息（个人详情页：基本信息 + 积分/配额 + 创作数量等统计）",
        "operationId": "get_user_profile_api_user_profile_get",
        "parameters": [
          {
            "name": "SESSION",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_UserProfileVO_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/list/page": {
      "post": {
        "tags": [
          "用户管理"
        ],
        "summary": "List Users By Page",
        "description": "分页查询用户列表（管理员）",
        "operationId": "list_users_by_page_api_user_list_page_post",
        "parameters": [
          {
            "name": "SESSION",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserQueryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/add": {
      "post": {
        "tags": [
          "用户管理"
        ],
        "summary": "Add User",
        "description": "添加用户（管理员）",
        "operationId": "add_user_api_user_add_post",
        "parameters": [
          {
            "name": "SESSION",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserAddRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_int_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/update": {
      "post": {
        "tags": [
          "用户管理"
        ],
        "summary": "Update User",
        "description": "更新用户（管理员）",
        "operationId": "update_user_api_user_update_post",
        "parameters": [
          {
            "name": "SESSION",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_bool_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/delete": {
      "post": {
        "tags": [
          "用户管理"
        ],
        "summary": "Delete User",
        "description": "删除用户（管理员）",
        "operationId": "delete_user_api_user_delete_post",
        "parameters": [
          {
            "name": "SESSION",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_bool_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/profile/update": {
      "post": {
        "tags": [
          "用户管理"
        ],
        "summary": "Update User Profile",
        "description": "更新当前登录用户的个人资料（昵称/头像/简介）\n\n更新成功后同步刷新 Redis Session 中的用户信息，\n保证下次 GET /user/get/login 返回最新资料。",
        "operationId": "update_user_profile_api_user_profile_update_post",
        "parameters": [
          {
            "name": "SESSION",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserProfileUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_LoginUserVO_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/change-password": {
      "post": {
        "tags": [
          "用户管理"
        ],
        "summary": "Change Password",
        "description": "修改当前登录用户的密码\n\n安全策略：修改成功后清除 Redis Session 与 Cookie，强制用户重新登录。",
        "operationId": "change_password_api_user_change_password_post",
        "parameters": [
          {
            "name": "SESSION",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserChangePasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_bool_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/avatar/upload": {
      "post": {
        "tags": [
          "用户管理"
        ],
        "summary": "Upload Avatar",
        "description": "上传用户头像（multipart/form-data，字段名 file）\n\n仅支持 JPG / PNG / WebP / GIF，大小不超过 2MB；\n文件保存到本地 static/images/avatar/，返回可访问的图片 URL。",
        "operationId": "upload_avatar_api_user_avatar_upload_post",
        "parameters": [
          {
            "name": "SESSION",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_avatar_api_user_avatar_upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_str_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BaseResponse_LoginUserVO_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "状态码",
            "default": 0
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LoginUserVO"
              },
              {
                "type": "null"
              }
            ],
            "description": "响应数据"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息",
            "default": "ok"
          }
        },
        "type": "object",
        "title": "BaseResponse[LoginUserVO]"
      },
...
      "BaseResponse_UserProfileVO_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "状态码",
            "default": 0
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserProfileVO"
              },
              {
                "type": "null"
              }
            ],
            "description": "响应数据"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息",
            "default": "ok"
          }
        },
        "type": "object",
        "title": "BaseResponse[UserProfileVO]"
      },
      "BaseResponse_UserVO_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "状态码",
            "default": 0
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserVO"
              },
              {
                "type": "null"
              }
            ],
            "description": "响应数据"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息",
            "default": "ok"
          }
        },
        "type": "object",
        "title": "BaseResponse[UserVO]"
      },
      "BaseResponse_bool_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "状态码",
            "default": 0
          },
          "data": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "响应数据"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息",
            "default": "ok"
          }
        },
        "type": "object",
        "title": "BaseResponse[bool]"
      },
      "BaseResponse_dict_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "状态码",
            "default": 0
          },
          "data": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "响应数据"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息",
            "default": "ok"
          }
        },
        "type": "object",
        "title": "BaseResponse[dict]"
      },
      "BaseResponse_int_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "状态码",
            "default": 0
          },
          "data": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "响应数据"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息",
            "default": "ok"
          }
        },
        "type": "object",
        "title": "BaseResponse[int]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UserAddRequest": {
        "properties": {
          "userAccount": {
            "type": "string",
            "title": "Useraccount",
            "description": "账号"
          },
          "userPassword": {
            "type": "string",
            "title": "Userpassword",
            "description": "密码"
          },
          "userName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username",
            "description": "用户昵称"
          },
          "userAvatar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Useravatar",
            "description": "用户头像"
          },
          "userProfile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Userprofile",
            "description": "用户简介"
          },
          "userRole": {
            "type": "string",
            "title": "Userrole",
            "description": "用户角色",
            "default": "user"
          }
        },
        "type": "object",
        "required": [
          "userAccount",
          "userPassword"
        ],
        "title": "UserAddRequest",
        "description": "添加用户请求（管理员）"
      },
...
      "UserLoginRequest": {
        "properties": {
          "userAccount": {
            "type": "string",
            "maxLength": 256,
            "minLength": 4,
            "title": "Useraccount",
            "description": "账号"
          },
          "userPassword": {
            "type": "string",
            "maxLength": 512,
            "minLength": 8,
            "title": "Userpassword",
            "description": "密码"
          }
        },
        "type": "object",
        "required": [
          "userAccount",
          "userPassword"
        ],
        "title": "UserLoginRequest",
        "description": "用户登录请求"
      },
...
      "UserRegisterRequest": {
        "properties": {
          "userAccount": {
            "type": "string",
            "maxLength": 256,
            "minLength": 4,
            "title": "Useraccount",
            "description": "账号"
          },
          "userPassword": {
            "type": "string",
            "maxLength": 512,
            "minLength": 8,
            "title": "Userpassword",
            "description": "密码"
          },
          "checkPassword": {
            "type": "string",
            "maxLength": 512,
            "minLength": 8,
            "title": "Checkpassword",
            "description": "确认密码"
          }
        },
        "type": "object",
        "required": [
          "userAccount",
          "userPassword",
          "checkPassword"
        ],
        "title": "UserRegisterRequest",
        "description": "用户注册请求"
      },
...
    }
  }
}